downstream_tasks/minja_annotation/requirements.txt
absl-py
accelerate==1.10.1
aiohappyeyeballs
aiohttp
aiosignal
antlr4-python3-runtime
anyio
appdirs
argcomplete
argh==0.31.3
arrow
asttokens
attrs
azure-core
azure-identity
azure-storage-blob
azure-storage-file-datalake
backoff
bcrypt
binaryornot
biopython
boto3
botocore
Brotli
cachetools
certifi
cffi
chardet
charset-normalizer
circuitbreaker
click
colorama
comm
contourpy
cookiecutter
coolname
cramjam
cryptography
cycler
dataclasses
datasets==2.21.0
debugpy
decorator
dill
docker-pycreds==0.4.0
einops
evaluate
exceptiongroup
execnet
executing
filelock
flash-attn==2.6.3
fonttools
frozenlist
fsspec
gffutils==0.13
gitdb
GitPython
gmpy2
google-api-core
googleapis-common-protos
google-auth
google-cloud-core
google-cloud-storage
google-crc32c
google-resumable-media
gql
graphql-core
grpcio
h2
h5py==3.14.0
hf-xet==1.1.10
hpack
huggingface-hub==0.34.4
hydra-core==1.3.2
hyperframe
idna
importlib_metadata
iniconfig
ipykernel
ipython
isodate
jedi
Jinja2
jmespath
jupyter_client
jupyter_core
kiwisolver
lightning-utilities
llvmlite==0.43.0
Markdown
markdown-it-py
MarkupSafe
matplotlib==3.9.2
matplotlib-inline
mdurl
mosaicml
mosaicml-cli==0.6.41
mosaicml-streaming==0.8.0
mpmath
msal-extensions
msal
multidict
multiprocess
munkres==1.1.4
nest_asyncio
networkx
ninja==1.11.1.1
numba
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvitop==1.5.1
oci
omegaconf
packaging
pandas
paramiko
parso
pathtools==0.1.2
pexpect
pickleshare
pillow
platformdirs
pluggy
portalocker
prompt-toolkit
protobuf==4.25.3
proto-plus
psutil
ptyprocess==0.7.0
pure_eval
pyarrow==17.0.0
pyarrow-hotfix
pyasn1
pyasn1_modules
pyBigWig==0.3.24
pycairo==1.26.1
pycparser
py-cpuinfo
pyfaidx==0.9.0.3
Pygments
PyGObject==3.48.2
PyJWT
PyNaCl
pyOpenSSL
pyparsing
pysam==0.23.3
PySide6==6.7.2
PySocks
pytest
pytest-xdist
python-dateutil
python-slugify
python-snappy
pytorch_ranger
pytz
pyu2f
PyYAML
pyzmq
questionary
regex
requests
responses
rich
rsa
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer
safetensors
sentry-sdk
setproctitle
shellingham
shiboken6==6.7.2
simplejson
six
smmap
sniffio
stack_data
sympy==1.13.1
tabulate
tensorboard_data_server==0.7.0
tensorboard==2.19.0
termcolor
text-unidecode
tokenizers==0.22.0
tomli
torch==2.6.0
torchaudio==2.6.0
torchmetrics==1.6.0
torch-optimi==0.2.1
torch-optimizer
torchvision==0.21.0
tornado
tqdm
traitlets
transformers==4.56.2
triton==3.2.0
typer==0.12.4
typer-slim==0.12.4
types-python-dateutil
typing_extensions
tzdata
urllib3
validators
wandb
wcwidth
websockets
Werkzeug
xxhash
yarl
zipp
zstandard==0.23.0
zstd==1.5.5.1
scipy